% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adelaide_theme.R
\name{adelaide_theme}
\alias{adelaide_theme}
\title{adelaide_theme}
\usage{
use w/ ggplot
}
\description{
adds a blank theme to an existing ggplot
}
\examples{
ggplot(data = msleep, aes(x = sleep_total, y = sleep_rem)) +geom_point() + THEME_NAME()
}
